Extracting the toolbar logic into an extension
